Move useStoreCart to base/context/hooks
