+++ assets/js/base/context/add-to-cart-form/form/submit/index.js
@@ -4,8 +4,7 @@
 import { __ } from '@wordpress/i18n';
 import triggerFetch from '@wordpress/api-fetch';
 import { useEffect, useCallback, useState } from '@wordpress/element';
-import { useStoreCart } from '@woocommerce/base-hooks';
-import { useStoreNotices } from '@woocommerce/base-context/hooks';
+import { useStoreCart, useStoreNotices } from '@woocommerce/base-context/hooks';
 import { decodeEntities } from '@wordpress/html-entities';
 import { triggerFragmentRefresh } from '@woocommerce/base-utils';
 
@@ -13,7 +12,7 @@ import { triggerFragmentRefresh } from '@woocommerce/base-utils';
  * Internal dependencies
  */
 import { useAddToCartFormContext } from '../../form-state';
-import { useValidationContext } from '../../../shared';
+import { useValidationContext } from '../../../validation';
 
 /**
  * FormSubmit.
