+++ assets/js/base/context/cart-checkout/checkout/processor/index.js
@@ -10,9 +10,8 @@ import {
 	useState,
 	useMemo,
 } from '@wordpress/element';
-import { useStoreCart } from '@woocommerce/base-hooks';
 import { formatStoreApiErrorMessage } from '@woocommerce/base-utils';
-import { useStoreNotices } from '@woocommerce/base-context/hooks';
+import { useStoreCart, useStoreNotices } from '@woocommerce/base-context/hooks';
 
 /**
  * Internal dependencies
@@ -22,7 +21,7 @@ import { useCheckoutContext } from '../../checkout-state';
 import { useShippingDataContext } from '../../shipping';
 import { useCustomerDataContext } from '../../customer';
 import { usePaymentMethodDataContext } from '../../payment-methods';
-import { useValidationContext } from '../../../shared';
+import { useValidationContext } from '../../../validation';
 
 /**
  * CheckoutProcessor component.
