+++ assets/js/base/context/cart-checkout/shipping/index.js
@@ -9,7 +9,8 @@ import {
 	useMemo,
 	useRef,
 } from '@wordpress/element';
-import { useStoreCart, useSelectShippingRates } from '@woocommerce/base-hooks';
+import { useSelectShippingRates } from '@woocommerce/base-hooks';
+import { useStoreCart } from '@woocommerce/base-context/hooks';
 import isShallowEqual from '@wordpress/is-shallow-equal';
 import { deriveSelectedShippingRates } from '@woocommerce/base-utils';
