+++ assets/js/base/context/hooks/cart/use-store-cart-coupons.ts
@@ -6,15 +6,15 @@
 import { __, sprintf } from '@wordpress/i18n';
 import { useSelect } from '@wordpress/data';
 import { CART_STORE_KEY as storeKey } from '@woocommerce/block-data';
-import { useValidationContext } from '@woocommerce/base-context';
 import { decodeEntities } from '@wordpress/html-entities';
 import type { StoreCartCoupon } from '@woocommerce/types';
-import { useStoreNotices } from '@woocommerce/base-context/hooks';
 
 /**
  * Internal dependencies
  */
 import { useStoreCart } from './use-store-cart';
+import { useStoreNotices } from '../use-store-notices';
+import { useValidationContext } from '../../validation';
 
 /**
  * This is a custom hook for loading the Store API /cart/coupons endpoint and an
