+++ assets/js/base/context/hooks/index.js
@@ -1,1 +1,2 @@
+export * from './cart';
 export * from './use-store-notices';
