+++ assets/js/base/context/shared/index.js
@@ -1,2 +1,1 @@
-export * from './validation';
 export * from './event-emit';
