+++ assets/js/base/hooks/index.js
@@ -1,4 +1,3 @@
-export * from './cart';
 export * from './checkout';
 export * from './customer';
 export * from './payment-methods';
