+++ assets/js/base/hooks/use-store-add-to-cart.ts
@@ -6,12 +6,7 @@ import { useDispatch } from '@wordpress/data';
 import { CART_STORE_KEY as storeKey } from '@woocommerce/block-data';
 import { decodeEntities } from '@wordpress/html-entities';
 import type { CartItem } from '@woocommerce/types';
-import { useStoreNotices } from '@woocommerce/base-context/hooks';
-
-/**
- * Internal dependencies
- */
-import { useStoreCart } from './cart';
+import { useStoreCart, useStoreNotices } from '@woocommerce/base-context/hooks';
 
 /**
  * @typedef {import('@woocommerce/type-defs/hooks').StoreCartItemAddToCart} StoreCartItemAddToCart
