+++ assets/js/blocks/cart-checkout/cart/full-cart/cart-line-item-row.js
@@ -7,10 +7,8 @@ import PropTypes from 'prop-types';
 import QuantitySelector from '@woocommerce/base-components/quantity-selector';
 import ProductPrice from '@woocommerce/base-components/product-price';
 import ProductName from '@woocommerce/base-components/product-name';
-import {
-	useStoreCartItemQuantity,
-	useStoreEvents,
-} from '@woocommerce/base-hooks';
+import { useStoreEvents } from '@woocommerce/base-hooks';
+import { useStoreCartItemQuantity } from '@woocommerce/base-context/hooks';
 import {
 	ProductBackorderBadge,
 	ProductImage,
