+++ assets/js/blocks/cart-checkout/cart/full-cart/index.js
@@ -22,8 +22,11 @@ import {
 	COUPONS_ENABLED,
 	DISPLAY_CART_PRICES_INCLUDING_TAX,
 } from '@woocommerce/block-settings';
-import { useStoreCartCoupons, useStoreCart } from '@woocommerce/base-hooks';
-import { useStoreNotices } from '@woocommerce/base-context/hooks';
+import {
+	useStoreCartCoupons,
+	useStoreCart,
+	useStoreNotices,
+} from '@woocommerce/base-context/hooks';
 import classnames from 'classnames';
 import {
 	Sidebar,
